refactor(renderer): extract the SSGI probe block into ssgi_pass.rs
Render-graph migration cluster 3 prep: the 360-line Lumen-style
screen-probe block (place / trace SW-HiZ-HW-SDF / temporal EMA /
octahedral resolve, plus the disabled-clear) becomes
record_ssgi_passes. One borrow reorder: ssr_composite_view now binds
after the SSGI call (compose reads it later either way). Goldens
pixel-identical. mod.rs ratchets 13464 -> 13110.
